backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
Markdown==3.7
sqlparse==0.5.3

# Nuove dipendenze per Sicurezza e JWT
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
dj-rest-auth==6.0.0
django-allauth==65.4.1

# Pagamenti
stripe==11.5.0

# Parsing e NLP
PyPDF2==3.0.1
python-docx==1.1.2
nltk==3.9.1
spacy==3.8.2

# Testing
pytest==8.3.4
pytest-django==4.9.0
coverage==7.6.1
factory-boy==3.3.1
